+++ llama-index-integrations/readers/llama-index-readers-file/llama_index/readers/file/__init__.py
@@ -12,7 +12,11 @@ from llama_index.readers.file.mbox import MboxReader
 from llama_index.readers.file.paged_csv import PagedCSVReader
 from llama_index.readers.file.pymu_pdf import PyMuPDFReader
 from llama_index.readers.file.slides import PptxReader
-from llama_index.readers.file.tabular import PandasCSVReader, CSVReader
+from llama_index.readers.file.tabular import (
+    PandasCSVReader,
+    CSVReader,
+    PandasExcelReader,
+)
 from llama_index.readers.file.unstructured import UnstructuredReader
 from llama_index.readers.file.video_audio import VideoAudioReader
 from llama_index.readers.file.xml import XMLReader
@@ -33,6 +37,7 @@ __all__ = [
     "MboxReader",
     "PptxReader",
     "PandasCSVReader",
+    "PandasExcelReader",
     "VideoAudioReader",
     "UnstructuredReader",
     "PyMuPDFReader",
