+++ llama-index-legacy/llama_index/legacy/readers/file/base.py
@@ -23,7 +23,6 @@ from llama_index.legacy.readers.file.mbox_reader import MboxReader
 from llama_index.legacy.readers.file.slides_reader import PptxReader
 from llama_index.legacy.readers.file.tabular_reader import (
     PandasCSVReader,
-    PandasExcelReader,
 )
 from llama_index.legacy.readers.file.video_audio_reader import VideoAudioReader
 from llama_index.legacy.schema import Document
@@ -45,8 +44,6 @@ DEFAULT_FILE_READER_CLS: Dict[str, Type[BaseReader]] = {
     ".md": MarkdownReader,
     ".mbox": MboxReader,
     ".ipynb": IPYNBReader,
-    ".xls": PandasExcelReader,
-    ".xlsx": PandasExcelReader,
 }
