+++ llama-index-legacy/llama_index/legacy/readers/file/tabular_reader.py
@@ -4,7 +4,6 @@ Contains parsers for tabular data files.
 
 """
 
-import importlib.util
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -115,98 +114,3 @@ class PandasCSVReader(BaseReader):
             return [
                 Document(text=text, metadata=extra_info or {}) for text in text_list
             ]
-
-
-class PandasExcelReader(BaseReader):
-    r"""Pandas-based Excel parser.
-
-    Parses Excel files using the Pandas `read_excel`function.
-    If special parameters are required, use the `pandas_config` dict.
-
-    Args:
-        concat_rows (bool): whether to concatenate all rows into one document.
-            If set to False, a Document will be created for each row.
-            True by default.
-
-        sheet_name (str | int | None): Defaults to None, for all sheets, otherwise pass a str or int to specify the sheet to read.
-
-        pandas_config (dict): Options for the `pandas.read_excel` function call.
-            Refer to https://pandas.pydata.org/docs/reference/api/pandas.read_excel.html
-            for more information.
-            Set to empty dict by default.
-
-    """
-
-    def __init__(
-        self,
-        *args: Any,
-        concat_rows: bool = True,
-        sheet_name=None,
-        pandas_config: dict = {},
-        **kwargs: Any
-    ) -> None:
-        """Init params."""
-        super().__init__(*args, **kwargs)
-        self._concat_rows = concat_rows
-        self._sheet_name = sheet_name
-        self._pandas_config = pandas_config
-
-    def load_data(
-        self, file: Path, extra_info: Optional[Dict] = None
-    ) -> List[Document]:
-        """Parse file."""
-        openpyxl_spec = importlib.util.find_spec("openpyxl")
-        if openpyxl_spec is not None:
-            pass
-        else:
-            raise ImportError(
-                "Please install openpyxl to read Excel files. You can install it with 'pip install openpyxl'"
-            )
-
-        # sheet_name of None is all sheets, otherwise indexing starts at 0
-        dfs = pd.read_excel(file, self._sheet_name, **self._pandas_config)
-
-        documents = []
-
-        # handle the case where only a single DataFrame is returned
-        if isinstance(dfs, pd.DataFrame):
-            df = dfs.fillna("")
-
-            # Convert DataFrame to list of rows
-            text_list = (
-                df.astype(str).apply(lambda row: " ".join(row.values), axis=1).tolist()
-            )
-
-            if self._concat_rows:
-                documents.append(
-                    Document(text="\n".join(text_list), metadata=extra_info or {})
-                )
-            else:
-                documents.append(
-                    [
-                        Document(text=text, metadata=extra_info or {})
-                        for text in text_list
-                    ]
-                )
-        else:
-            for df in dfs.values():
-                df = df.fillna("")
-
-                # Convert DataFrame to list of rows
-                text_list = (
-                    df.astype(str).apply(lambda row: " ".join(row), axis=1).tolist()
-                )
-
-                if self._concat_rows:
-                    documents.append(
-                        Document(text="\n".join(text_list), metadata=extra_info or {})
-                    )
-                else:
-                    documents.append(
-                        [
-                            Document(text=text, metadata=extra_info or {})
-                            for text in text_list
-                        ]
-                    )
-
-        return documents
